Fix 64-bit audio bugs and crashes related to gyroids (haniwa)
This commit resolves two major bugs in the 64-bit build:
1. Gyroids being "stuck" (no animation or sound) upon entering a house.
2. The game crashing when stopping and restarting a gyroid.

Root Causes:
- Pointer truncation: Object pointers were used as IDs in the rhythm and
  audio systems, being cast to `u32`. This caused ID collisions and
  incorrect state lookups on 64-bit systems.
- Stack corruption: `Z_osRecvMesg` was receiving pointer-sized messages
  into 32-bit variables, corrupting the stack.
- RSP address space: The software RSP emulator and `Acmd` structure
  were limited to 32-bit addresses, causing memory access violations.

Key Fixes:
- Transitioned audio IDs and internal structures (`NA_RHYTHM_BUFFER`,
  `SOU_ONGEN_ENTRY`, etc.) to use `uintptr_t`.
- Standardized message queues to use `OSMesg` for all transfers.
- Modernized `Acmd` and `rspsim.c` to support 64-bit host pointers
  natively, allowing the removal of legacy address recovery hacks.
- Widened DMA and relocation arithmetic to support 64-bit memory layouts.
- Corrected various pointer-to-integer casts in sequence tracking and
  bank loading logic.

+++ include/dolphin/ar.h
@@ -38,7 +38,7 @@ struct ARQRequest {
 // ARQ functions.
 void ARQInit();
 #ifdef TARGET_PC
-void ARQPostRequest(void* req, u32 owner, u32 type, u32 prio, uintptr_t source, uintptr_t dest, u32 length, void* callback);
+void ARQPostRequest(void* req, uintptr_t owner, u32 type, u32 prio, uintptr_t source, uintptr_t dest, u32 length, void* callback);
 #else
 void ARQPostRequest(ARQRequest* task, u32 owner, u32 type, u32 priority, u32 source, u32 dest, u32 length, ARQCallback callback);
 #endif
